api/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-storage-queue==12.11.0
requests==2.32.3
